requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
billiard==4.2.0
blessed==1.20.0
celery==5.3.6
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==1.4.0
Django==4.2.9
django-celery==3.1.17
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.5
django-fsm==2.8.1
django-picklefield==3.1
django-q==1.3.9
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.4
gunicorn==23.0.0
h11==0.14.0
honcho==1.1.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.4
packaging==23.2
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
requests==2.32.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.27.0
vine==5.1.0
wcwidth==0.2.13
